Merging validation and testing branch

+++ services/paymentGateway/src/main/java/org/bazaar/paymentGateway/payment/entity/UserAccount.java
@@ -1,5 +1,7 @@
 package org.bazaar.paymentGateway.payment.entity;
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -21,5 +23,5 @@ public class UserAccount {
     @Column(unique = true)
     private String email;
     private String password;
-    private float moneyInAccount;
+    private BigDecimal moneyInAccount;
 }
